fix: harden cache semantics, centralized log throttling, and worker broadcast isolation
The new LogThrottle utility replaces the hand-rolled volatile check-then-act timestamp windows previously copied across VersionControllerImpl, WorkerHeartbeatProducer, WorkerBroadcastBuffer, TransitionLogThrottle, and CacheExtensionAspect. It uses an AtomicLong compare-and-set so exactly one caller per window is admitted, seeds the first call to always win, and measures elapsed time through TimeSource.monotonicMillis so a backward clock step cannot silently suppress an alarm. Its Counting and Accumulator variants cover the “one full WARN per window, tally the rest” and aggregate-summary idioms without each call site reimplementing the counter and reset dance.

The annotation cache path now preserves the facade read contract that every read triggers detection and reporting. ZetaSpringCache precomputes the cache key prefix once, routes lookup through peekAndTag instead of a separate peek plus sentinel inspection, passes the caller policy into computeIfAbsentWithSoftExpire via withReader, and makes put honor the resolved hard and soft TTLs, null-caching, NullValue sentinel translation, and skipBroadcast. A null with null-caching disabled writes nothing, while a cached null uses the short null TTL and still surfaces Spring’s marker. clear now invalidates only keys carrying this cache’s prefix through one batched invalidate rather than wiping the entire L1. CacheExtensionAspect captures the key separator at construction, caches parameter names, rethrows ZetaBlockedException without consulting a fallback, builds a StandardEvaluationContext with p/a aliases plus discovered parameter names, rate-limits fallback, cache-condition, TTL-SpEL, and preload-SpEL warnings, and bounds the preload count to Integer.MAX_VALUE so re-registration after dedup expiry cannot overflow the detector.

The soft-TTL disabled contract is corrected in TtlPolicy: softExpireAtMs of zero now reports not soft-expired, since zero is the documented disabled value and the old “immediately expired” reading armed background refreshes on every hit. TransactionSupport now absorbs submit-time RejectedExecutionException and wraps the afterCommit task body so a deferred cache update failure cannot turn an already-committed transaction into a caller-visible error. ZetaExceptionHandler gained a fallbackToWarn overload for call sites that own their own logging, ZetaBlockedException is constructed without a stack trace for high-QPS blocked keys, NullValue gained a serialVersionUID, and ZetaModeException hoisted its source-class constant.

Worker and messaging changes add ADR-0068 application isolation: broadcasts carry the sender’s appName header and receivers are documented to drop foreign-app decisions, with a matching HEADER_ORIGIN_INSTANCE constant for ADR-0067 sync messages. WorkerMessage delegates header parsing to the shared AmqpMessageReader, queue-name prefixes moved into ZetaConstants.Routing, and the heartbeat RabbitTemplate pre-declares the control-plane exchange so a cold cluster start does not publish to a not-yet-existing exchange and lose the first heartbeat. WorkerHeartbeatProducer uses LogThrottle and drops the explicit @DependsOn rabbitAdmin, WorkerHeartbeatVerifier synchronizes stop and emits at most one cluster-health WARN per verification round, WorkerProperties adds @Max(60000) on rebroadcastIntervalMs, and WorkerAutoConfiguration validates that interval at startup. WorkerBroadcastBuffer also switches its aggregated failure WARN to the shared throttle.

Configuration and documentation received supporting updates: ZetaAutoConfiguration passes the power-of-two alignment flag into HeavyKeeper, ZetaProperties exposes DEFAULT_APP_NAME, ZetaLockAutoConfiguration orders itself after RedisAutoConfiguration, ZetaEndpoint documents /actuator/hotkey and now reports msSinceLastAnyHeartbeat, and the MONITOR documents the sketch-width power-of-two alignment and the config-negotiation validation invariant. The .gitignore and IntelliJ metadata files were adjusted as housekeeping.

Tests and docs were updated accordingly. LogThrottleTest covers first-call, window suppression, exact boundary, concurrent burst admission, constructor validation, and independent instances. BroadcastBufferMergeTest pins version-ordered merge behavior. TtlPolicyTest now expects zero soft-expire to be disabled. ZetaCacheTest gains extensive coverage for lazy TTL suppliers, REVALIDATE keep-and-renew, null-sentinel penetration protection, batch sentinel parity, CAS and compareAndInvalidate latched races, putIfAbsent/getAndSet dedup invalidation, tag normalization, and peekAndTag whitelist behavior. ZetaSpringCacheTest adds null-caching TTL, clear isolation, and integration over a real HotKeyCache. WorkerAutoConfigurationTest covers rebroadcast-interval bounds, WorkerBroadcasterTest asserts the appName header, ZetaReporterTest verifies routing-task failure accounting, and TransactionSupportTest covers executor rejection. The ZetaCacheExtensionAspectTest updates preload counts from Long.MAX_VALUE to Integer.MAX_VALUE.

+++ common/src/test/java/io/github/hyshmily/zeta/annotation/ZetaCacheExtensionAspectTest.java
@@ -1152,7 +1152,10 @@ class ZetaCacheExtensionAspectTest {
     aspect.aroundCacheable(pjp);
 
     verify(zeta).notifyLocalDetectorDirect(
-      Map.of("test::preload-key-a", Long.MAX_VALUE, "test::preload-key-b", Long.MAX_VALUE)
+      Map.of(
+        "test::preload-key-a", (long) Integer.MAX_VALUE,
+        "test::preload-key-b", (long) Integer.MAX_VALUE
+      )
     );
   }
 
@@ -1175,7 +1178,7 @@ class ZetaCacheExtensionAspectTest {
 
     aspect.aroundCacheable(pjp);
 
-    verify(zeta).notifyLocalDetectorDirect("test::myDynamicKey", Long.MAX_VALUE);
+    verify(zeta).notifyLocalDetectorDirect("test::myDynamicKey", Integer.MAX_VALUE);
   }
 
   @Test
@@ -1200,7 +1203,10 @@ class ZetaCacheExtensionAspectTest {
     aspect.aroundCacheable(pjp);
 
     verify(zeta, times(1)).notifyLocalDetectorDirect(
-      Map.of("test::preload-key-a", Long.MAX_VALUE, "test::preload-key-b", Long.MAX_VALUE)
+      Map.of(
+        "test::preload-key-a", (long) Integer.MAX_VALUE,
+        "test::preload-key-b", (long) Integer.MAX_VALUE
+      )
     );
   }
 }
